police cars with smart AI

+++ Classes/Obstacle.cpp
@@ -1,5 +1,5 @@
 #include "Obstacle.h"
 
-Obstacle::Obstacle(std::string texture, int x, int y, int width, int height) : GameObject(texture, x, y, width, height) {
+Obstacle::Obstacle(std::string texture, float x, float y, float width, float height) : GameObject(texture, x, y, width, height) {
 
 }
